renamed controller and service files for loanhistory
