+++ backend/routes/controllers/loanhistory_controller.js
@@ -1,7 +1,7 @@
 const auth = require('../../utils/auth');
 const connection = require('../../utils/connection');
 
-const loanhistory = async (req, res, next) => {
+const loanhistory_controller = async (req, res, next) => {
     const {cust_id} = auth.decode_jwt(req.headers.access_token);
     const sql = "SELECT customerloan.CustomerId,Loan.LoanId,Loan.loan_amount FROM customer NATURAL JOIN customerloan NATURAL JOIN loan WHERE customer.CustomerId=?";
     const con = await connection();
@@ -11,4 +11,4 @@ const loanhistory = async (req, res, next) => {
 
 }
 
-module.exports = loanhistory;
+module.exports = loanhistory_controller;
