+++ backend/routes/services/loanhistory_service.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const auth = require('../../utils/auth');
-const loanhistory = require('../controllers/loanhistory');
+const loanhistory_service = require('../controllers/loanhistory_controller');
 const {log} = require("debug");
 
-router.get('/', auth.auth_jwt, loanhistory);
+router.get('/', auth.auth_jwt, loanhistory_service);
 
 module.exports = router;
